requirements.txt
future
numpy==1.18.4
pygame
matplotlib
open3d
Pillow
psutil
py-cpuinfo
pygame
python-tr
networkx
distro
opencv-python
pandas